Release iteration #3 (#48)
* EBOCK-125_TechDebt-132-133-137

* Ebock 141 dashboard admin gestion des utilisateurs (#27)

* Ajout lister utilisateurs

* Ajout enable + disable

* Ajout test admin

* Ajout test

* Ajout validation cip

* Fix commentaire PR

* Changement paths

* Fix test enable

* Ajout delete aux specifications (#28)

* Ajout delete aux specifications

* Fix dislexie

* Remove useless import

* EBOCK-35_AjouterUneRevueAUnVendeur (#29)

* ajout du payload, ajout service et dans les mapper

reste à faire les 2 query dans les mapper.xml

* ajout nouveau mapper (update et getCountByUsers) et tous les mapper.xml

* ajout tests service

* ajout classe ReviewServiceIT

* ajout vérification d'une string

* ajout champ update

* changement nom

* quick fix

* Response -> throw

* quick fix

* Changed README

* Ebock 162 item update put doit pouvoir etre partiel (#30)

* Item partial update

* Fix test item update

* EBOCK-125_TechDebt-136 (#32)

* EBOCK-125_TechDebt_128_135 (#33)

* Ajout city to address

* Remove warning + mini fix city

* Throw 400 on FK violation

* EBoCK-139_Changer_photo_profile (#34)

* Ajout editProfilePicture

* Fix test

* Fix test EditProfilePicture

* EBOCK-173_Fix_Enable_Disable_Compte_Admin (#36)

* Fix enable/disable user

* Fix edit user profile picture

* EBOCK-34_GestionDesCommentairesSurUnProduit (#37)

* modif de review

insert retourne 201

* ajout au dernier commit

* payload pour insert et response pour les détails des commentaires

* les mappers et le service

details, insert, delete

* ajout des tests

* modif tests review

* modif service

* modif tests

* changement length max de content pour review et comment

* modif des services item et comment

insert et details de comment sont rendus dans itemService (pour id)

* modif des tests

* changement nom

* quick fix fucky-wucky

CommentService et les tests

* retourne 404 quand delete comment inexistant

* Revert "retourne 404 quand delete comment inexistant"

This reverts commit 3e75f2e9a08ece7861504873314da8618902d789.

* EBOCK-125_TechDebt-175 (#39)

* Fix test jenkuis

* Revert "Fix test jenkuis"

This reverts commit 712e7d14214ffb5b2034c725cee0fd78635b923f.

* EBOCK-121_ArchivageDUnItemConversation (#41)

* changenent min value quantity item

* changement query qui get all rooms

ne montre pas les rooms archived

* changments pour enlever archived

* service/mapper ajout delete et archive les convo (aussi quand qté 0)

aussi modif archived dans mapper item

* modif service

* tests

* modif dans les tests

* Update schema.sql

* ajout condition pour ne pas pouvoir modifier un item déjà archivé

* throws forbidden quand item déjà à 0

* EBOCK-121-139_Archivage_ProfilePic_FixRandomStuff (#42)

* fix random bullshit

* can remove profile picture

* revert globalExceptionMapper change

* EBOCK-94_Ajout_dark_ebock (#40)

* Ajout dark ebock

* OpenApi qui marche enfin

* Ajout patch

* Ajout addDark + removeDark

* Fix test jenquins

* Changement output erreur du [REDACTED]

* Fix schema.sql

* Fix schema.sql part 2

* quick item insert fix shhh

* EBOCK-185_BugFixWebsocket (#43)

* EBOCK-184_AjouterPhotoProfilCommentaireMessage (#44)

* Comment profile picture functionnal

* message and room profile picture

* Fix réactivé spécifications (#45)

* EBOCK-178_CommunicationDansUneConversationArchivee (#46)

* ajout du endpoint pour archiver une conversation

* ajout désarchive les convo quand acheteur écrit

* changement query getAllRooms

acheteur voit toutes les convo (même archived)
vendeur voit juste ceux pas archived

* ajout des tests et petit changement

* Update schema.sql

* EBOCK-95_AllouerAccesRedacted (#47)

* change header cors authorizations

* adress error fixes

* removed unconditional ebock. in mappers

* Added environment condition on websocket

---------


---------

+++ src/main/java/com/ebock/service/ItemService.java
@@ -2,8 +2,11 @@ package com.ebock.service;
 
 import com.ebock.business.Item;
 import com.ebock.converter.ItemConverter;
-import com.ebock.dto.request.item.FilterItemPayload;
-import com.ebock.dto.request.item.ItemPayload;
+import com.ebock.dto.request.comment.CommentPayload;
+import com.ebock.dto.request.item.FilterItemParameters;
+import com.ebock.dto.request.item.ItemCreatePayload;
+import com.ebock.dto.request.item.ItemUpdatePayload;
+import com.ebock.dto.response.comment.CommentDetailsResponse;
 import com.ebock.dto.response.item.ItemDetailsResponse;
 import com.ebock.dto.response.item.ItemInsertResponse;
 import com.ebock.dto.response.item.ItemResponse;
@@ -11,6 +14,7 @@ import com.ebock.mapper.*;
 import io.quarkus.security.Authenticated;
 import io.quarkus.security.UnauthorizedException;
 import jakarta.annotation.security.PermitAll;
+import jakarta.annotation.security.RolesAllowed;
 import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
@@ -44,6 +48,8 @@ public class ItemService {
     UserMapper userMapper;
     @Inject
     ItemConverter itemConverter;
+    @Inject
+    CommentMapper commentMapper;
     @Context
     SecurityContext securityContext;
 
@@ -61,16 +67,16 @@ public class ItemService {
                                    @RestQuery @Separator(",") List<Integer> wears,
                                    @RestQuery @Separator(",") List<Integer> deliveries,
                                    @RestQuery @Separator(",") List<Integer> payments) {
-        FilterItemPayload filterItemPayload = new FilterItemPayload();
-        filterItemPayload.minPrice = minP;
-        filterItemPayload.maxPrice = maxP;
-        filterItemPayload.maxDistance = maxD;
-        filterItemPayload.favorite = fav;
-        filterItemPayload.listCategoryId = categories;
-        filterItemPayload.listTagId = tags;
-        filterItemPayload.listWearId = wears;
-        filterItemPayload.listDeliveryId = deliveries;
-        filterItemPayload.listPaymentId = payments;
+        FilterItemParameters filterItemParameters = new FilterItemParameters();
+        filterItemParameters.minPrice = minP;
+        filterItemParameters.maxPrice = maxP;
+        filterItemParameters.maxDistance = maxD;
+        filterItemParameters.favorite = fav;
+        filterItemParameters.listCategoryId = categories;
+        filterItemParameters.listTagId = tags;
+        filterItemParameters.listWearId = wears;
+        filterItemParameters.listDeliveryId = deliveries;
+        filterItemParameters.listPaymentId = payments;
         int pageSize = 25;
 
         if (pageNumber < 1) {
@@ -81,7 +87,7 @@ public class ItemService {
             cip = securityContext.getUserPrincipal().getName();
         } catch (Exception e){}
 
-        return this.itemMapper.getPaginatedItem(pageNumber, pageSize, filterItemPayload, cip);
+        return this.itemMapper.getPaginatedItem(pageNumber, pageSize, filterItemParameters, cip);
     }
 
     @GET
@@ -131,10 +137,10 @@ public class ItemService {
     }
 
     @POST
-    @Path("/insert")
+    @Path("")
     @Authenticated
     @Transactional
-    public ItemInsertResponse insert(@Valid ItemPayload itemInsertPayload){
+    public ItemInsertResponse insert(@Valid ItemCreatePayload itemInsertPayload){
         Item item = itemConverter.toBusiness(itemInsertPayload);
         String cip = securityContext.getUserPrincipal().getName();
         item.sellerCip = cip;
@@ -161,11 +167,11 @@ public class ItemService {
     }
 
     @PUT
-    @Path("/update/{id}")
+    @Path("/{id}")
     @Authenticated
     @Transactional
-    public ItemInsertResponse update(@PathParam("id") int itemId, @Valid ItemPayload itemInsertPayload){
-        Item item = itemConverter.toBusiness(itemInsertPayload);
+    public Response update(@PathParam("id") int itemId, @Valid ItemUpdatePayload itemUpdatePayload){
+        Item item = itemConverter.toBusiness(itemUpdatePayload);
         item.itemId = itemId;
 
         String cip = securityContext.getUserPrincipal().getName();
@@ -179,32 +185,83 @@ public class ItemService {
             throw new ForbiddenException("Not your item");
         }
 
+        if(existingItem.quantity == 0) {
+            throw new ForbiddenException("Cannot modify item out of stock");
+        }
+
+        if (item.quantity == 0) {
+            itemMapper.archiveRoomsById(item.itemId);
+        }
+
         itemMapper.update(cip, item);
 
         // Update tags
-        itemTagMapper.deleteByItemId(itemId);
-        if (itemInsertPayload.tagList != null && !itemInsertPayload.tagList.isEmpty()) {
-            itemTagMapper.insert(item.itemId, itemInsertPayload.tagList);
+        if (itemUpdatePayload.tagList != null) {
+            itemTagMapper.deleteByItemId(itemId);
+            if (!itemUpdatePayload.tagList.isEmpty()) {
+                itemTagMapper.insert(item.itemId, itemUpdatePayload.tagList);
+            }
         }
 
         // Update images
-        itemImageMapper.deleteByItemId(itemId);
-        if(itemInsertPayload.imageList != null && !itemInsertPayload.imageList.isEmpty()){
-            itemImageMapper.insert(item.itemId, itemInsertPayload.imageList);
+        if (itemUpdatePayload.imageList != null) {
+            itemImageMapper.deleteByItemId(itemId);
+            if (!itemUpdatePayload.imageList.isEmpty()) {
+                itemImageMapper.insert(item.itemId, itemUpdatePayload.imageList);
+            }
         }
 
         // Update payment option
-        itemPaymentOptionMapper.deleteByItemId(itemId);
-        if (itemInsertPayload.paymentOptionList != null && !itemInsertPayload.paymentOptionList.isEmpty()) {
-            itemPaymentOptionMapper.insert(item.itemId, itemInsertPayload.paymentOptionList);
+        if (itemUpdatePayload.paymentOptionList != null) {
+            itemPaymentOptionMapper.deleteByItemId(itemId);
+            if (!itemUpdatePayload.paymentOptionList.isEmpty()) {
+                itemPaymentOptionMapper.insert(item.itemId, itemUpdatePayload.paymentOptionList);
+            }
         }
 
         // Update delivery option
-        itemDeliveryOptionMapper.deleteByItemId(itemId);
-        if (itemInsertPayload.deliveryOptionList != null && !itemInsertPayload.deliveryOptionList.isEmpty()) {
-            itemDeliveryOptionMapper.insert(item.itemId, itemInsertPayload.deliveryOptionList);
+        if (itemUpdatePayload.deliveryOptionList != null) {
+            itemDeliveryOptionMapper.deleteByItemId(itemId);
+            if (!itemUpdatePayload.deliveryOptionList.isEmpty()) {
+                itemDeliveryOptionMapper.insert(item.itemId, itemUpdatePayload.deliveryOptionList);
+            }
         }
 
-        return itemConverter.toInsertResponse(item);
+        return Response.noContent().build();
+    }
+
+    @GET
+    @Path("/{id}/comment")
+    @PermitAll
+    public List<CommentDetailsResponse> listItemComments(@PathParam("id") Integer id) {
+        if(itemMapper.getItemCountById(id) == 0)
+            throw new NotFoundException("Item not found");
+
+        return commentMapper.getDetailledComments(id);
+    }
+
+    @POST
+    @Path("/{id}/comment")
+    @Authenticated
+    public Response insertComment(@PathParam("id") Integer id, @Valid CommentPayload commentPayload){
+        if(itemMapper.getItemCountById(id) == 0)
+            throw new NotFoundException("Item not found");
+
+        String cip = securityContext.getUserPrincipal().getName();
+
+        commentMapper.insert(id, cip, commentPayload);
+        return Response.status(Response.Status.CREATED).build();
+    }
+
+    @DELETE
+    @Path("/{id}")
+    @RolesAllowed("admin")
+    public Response delete(@PathParam("id") int id){
+        itemMapper.delete(id);
+        int rowsAffected = itemMapper.archiveRoomsById(id);
+
+        if(rowsAffected == 0 && itemMapper.getItemCountById(id) == 0)
+            throw new NotFoundException("Item not found");
+        return Response.noContent().build();
     }
 }
